fixed npe for unsaved blocks

+++ src/main/java/de/raidcraft/RaidCraftPlugin.java
@@ -103,7 +103,9 @@ public class RaidCraftPlugin extends BasePlugin implements Component {
         }
         PlayerPlacedBlock playerPlacedBlock = new PlayerPlacedBlock(block);
         PlayerPlacedBlock remove = playerPlacedBlocks.remove(playerPlacedBlock);
-        getDatabase().delete(remove);
+        if (remove.getId() > 1) {
+            getDatabase().delete(remove);
+        }
     }
 
 }
